requirements.txt
appnope==0.1.4
asgiref==3.7.2
asttokens==2.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
decorator==5.1.1
Django==4.2.10
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-webpack-loader==3.0.1
djangorestframework==3.14.0
exceptiongroup==1.2.0
executing==2.0.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.1
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pytest==8.0.2
pytest-django==4.8.0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.1
typing_extensions==4.10.0
wcwidth==0.2.13
